+++ src/main/kotlin/com/procurement/notice/model/ocds/Value.kt
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.annotation.JsonPropertyOrder
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
 import com.procurement.notice.databinding.MoneyDeserializer
+import java.math.BigDecimal
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder("amount", "currency")
@@ -12,7 +13,7 @@ data class Value(
 
         @JsonProperty("amount")
         @JsonDeserialize(using = MoneyDeserializer::class)
-        var amount: Double?,
+        var amount: BigDecimal?,
 
         @JsonProperty("currency")
         val currency: Currency?
